Refine packet summary parsing
